style(phpcs): move RestoresDeletedAncestors out of Resource namespace
Connector actions still map Resource/Delete to Processors\Resource, so
that package cannot be renamed. Exclude the soft-reserved keyword sniff
there and in the legacy processor tests.
